requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
Pillow==10.1.0
python-dotenv==1.0.0
email-validator==2.1.0
